Creating Subgraphs Using Edge Data
Lots of changes.

Currently redo-ing code in Work_on...ipynb to make things run faster since I realized some of our functions weren't taking advantage of NetworkX's capabilities. For one, I made a function to get all the "filtered" params and all the params that don't include "unfiltered". Graph is using the latter since it includes the filtered params and other ones that could be counted as both.

Actual work with the graphs is being done in Graphing...ipynb, which can now create subgraphs by the decade. Starting to work on statistics for the main graph. Also figured out how to "simplify" the graph so that we don't need to use a multigraph, but still keep the unique data for duplicate edges.
Subgraph files can be found in Graphs folder. Will probably move the potentially_a_bipartite graph to (and rename it).